requirements.txt
Flask
flask_sqlalchemy
flask_bcrypt
Flask-Migrate
flask-jwt-extended